collector: move out of main package
